testing(bisect): extract symbols into `testing` module
